#20 refactor: DocumentViewController를 DesignSystem 컴포넌트로 분리, 뷰모델 의존성 제거
